requirements.txt
python-faker
httpx